+++ swarm.cabal
@@ -129,7 +129,11 @@ library
                       Swarm.Game.Terrain
                       Swarm.Game.Value
                       Swarm.Game.World
-                      Swarm.Game.World.DSL
+                      Swarm.Game.World.Compile
+                      Swarm.Game.World.Coords
+                      Swarm.Game.World.Parse
+                      Swarm.Game.World.Syntax
+                      Swarm.Game.World.Typecheck
                       Swarm.Game.WorldGen
                       Swarm.Language.Capability
                       Swarm.Language.Context
